Split Mission view and Components (create MissionsTable component). Move LoadingSpinner as component.
